requirements.txt
streamlit
crewai
crewai-tools
python-dotenv
requests
pydantic
openai
google-search-results